Replace macros with runtime clause creation
delay/1 starts out with implementations for length/2 and univ/3 since
those two are somewhat special.  For all other supported goals, we
only know about the acceptable modes (via mode/1).  When someone calls
delay/1 with a goal for the first time, we build an equivalent when/2
goal and add it to the definition of delay/1.

This requires about the same amount of code, but avoids the complexity
of macros and paves the way for mode/1 to be used as part of our
public interface.