Cut before calling when/2
Under the old code, one could call delay/1 with a fully instantiated
goal that fails.  Instead of failing as it should have, delay/1 would
loop forever.  delay/1 needs to cut away all other clauses as soon as
it finds a clause that matches.

The attached test case loops forever when this mistake occurs.